LayoutTests/platform/chromium-mac/fast/forms/input-list-expected.txt

Test for the list attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL input.list should be null (of type object). Was undefined (of type undefined).
FAIL input.list should be null (of type object). Was undefined (of type undefined).
FAIL input.list should be null (of type object). Was undefined (of type undefined).
FAIL input.list should be [object HTMLUnknownElement] (of type object). Was undefined (of type undefined).
FAIL input.list should be null (of type object). Was undefined (of type undefined).
FAIL datalist.className should be former. Threw exception TypeError: Cannot read property 'className' of undefined
PASS successfullyParsed is true

TEST COMPLETE
         
